clustering code

+++ snipgenie/tables.py
@@ -751,7 +751,7 @@ class SampleTableModel(DataFrameModel):
             elif type(value) != str:
                 if type(value) in [float,np.float64] and np.isnan(value):
                     return ''
-                elif type(value) == np.float:
+                elif type(value) == np.float32:
                     return value
                 else:
                     return (str(value))
